ohdsi_icd_elix

For extracting ICD codes from OMOP CDM to calculate Elixhauser comorbidity scores

ohdsi_icd_elix

Format

A data frame with 5 variables: concept_id, concept_name, vocabulary_id, concept_code, concept_code_strip

Source

Elixhauser, A., Steiner, C., Harris, D. R. & Coffey, R. M. Comorbidity measures for use with administrative data. Med. Care 36, 8–27 (1998)

van Walraven C, Austin PC, Jennings A, Quan H and Forster AJ. A modification of the Elixhauser comorbidity measures into a point system for hospital death using administrative data. Medical Care 2009; 47(6):626-633. DOI: 10.1097/MLR.0b013e31819432e5