Clean and process EHR data for linear mixed model analysis

process_clinical_data_lmm clean and process EHR data for linear mixed model analysis.

process_clinical_data_lmm(clinical_results, r_i)

Arguments

clinical_results
A tibble

r_i
A tibble

Details

This function cleans and processes clinical data from electronic health record (EHR) databases that have adopted the Observational Medical Outcomes Partnership (OMOP) Common Data Model (CDM) format, for linear mixed model analysis.

Updated: 2021-10-10